Fix potential use of uninitialised stack variable
Issue reported by Paul Rubin.

In the UTF-8 decoder, an unexpected continuation character may lead to
an uninitialised stack variable being used as the character code.
It's probably not exploitable so I'm doing this as a normal commit
rather than a security release. Fixed by initialising the variable.

Strengthened the documentation warnings about prevalidation of the input.

Change-Id: I01754dc6c2ffa9a94bbfb0f3d377f381a479b8c2