Add more test cases for recently changed diffing behavior
Source for all these is
https://wmde-wikidiff2-patched.wmflabs.org/core/index.php/Main_Page

We are adding these tests now to make sure later changes (like the
current bugfix in T197157) don't introduce undetected regressions. I
manually confirmed this is the case. Tests 09 and 10 got added
especially to fail without the bugfix from T197157. Test 11 also fails
because of the same bug, even if the test case originates from T195373.
I confirmed this is correct.

Bug: T201914
Change-Id: I5022b6d337dd61daf28b5479dcb700e1545c68d2