Keep break positions in a vector, rather than a set
explodeWords() finds break positions in a string by walking through it
once (common case) or twice (if any Thai characters are found). The
break positions found by each pass are guaranteed to be increasing and
unique, which makes an std::set unnecessary -- we can keep them as two
sorted vector subsequences, and then merge / deduplicate. This results
in a ~10% win on diffing chinese-reverse-{1,2}.php.

Change-Id: I47829c9e7a0606111fef564cb08968c791b1fb5f