Add test for T217546
This bug behaves pretty much like T215293 but the fix differs slightly
and needs a seperate test since #16 would not trigger it.

The example wiki text is a slighly modified version from the source diff
but contains the lines relevant to trigger the issue.

Also slighly changes the comment in test #16.

Bug: T217546
Change-Id: I7b95f69eb8d4d70a6b0d76b1939c9da0a0889dc7