Use a bitset+unordered_set to represent sets of integers
Change IntSet from an std::set to a custom container class that uses a
bitset to store small values and falls back to an std::unordered_set for
values that exceed the bitset size. This exploits the fact that we use
IntSet to store line numbers, which are typically small.

The bitset size is fixed at 512 bytes, allowing it to hold values in the
range [0, 4096).

benchmarks/bench.php results before/after on a 3.6Ghz Broadwell CPU:

Before:
 [en] best:  36.98ms   worst:  43.25ms   avg:  38.13ms
 [hi] best:  94.01ms   worst: 105.59ms   avg:  97.98ms
 [ru] best:  80.00ms   worst:  85.17ms   avg:  82.33ms
 [th] best:  27.52ms   worst:  31.24ms   avg:  28.38ms
 [zh] best:  25.34ms   worst:  33.30ms   avg:  26.49ms

After:
 [en] best:  27.36ms   worst:  29.73ms   avg:  28.15ms  (-26.17%)
 [hi] best:  40.36ms   worst:  41.95ms   avg:  41.04ms  (-58.11%)
 [ru] best:  54.21ms   worst:  57.28ms   avg:  55.20ms  (-32.95%)
 [th] best:  26.12ms   worst:  30.12ms   avg:  26.91ms  ( -5.18%)
 [zh] best:  18.87ms   worst:  20.51ms   avg:  19.26ms  (-27.29%)

The worst-case benchmark (chinese-reverse-bench.php) shows a small but
consistent win (-3%).

Change-Id: I326ef1ed3665c488682be7e5e5a469ad083b5f71