Add more benchmarks
Add a benchmark script that times diff generation for 50 changes,
sampled from Special:RecentChanges of several Wikipedias (en, hi, ru,
th, and zh).

There are a number of reasons why this workload may not be
representative, but it's still a better test of "typical" performance
than chinese-reverse-bench.php, which looks at a worst-case example.

Change-Id: I2db5e18d1fe8c9d7ae91a68464ba0a4461d9a91d