Return section titles in JSON formatter
This change attempts to return the related section title for each diff line in the JSON formatter.
It expects section title byte offsets to be passed in from outside. Originally section titles were
detected internally via regex, but it was too fragile.
Note only section titles for lines in the diff output are returned, not all section titles.

Bug: T234096
Change-Id: I5f5628505af6b3e0fdfe73f37f0d4bedc199a123