Don't use const for value parameters
Maybe there was some principled reason for this, but it looks like it
was just by analogy with const reference parameters. It doesn't really
make sense for the interface to place constraints on the constness of
local variables within a function implementation. In any case, it's
inconsistent from a style perspective.

Change-Id: I150d3d145587af10a079d513cb960324c520a2af