Remove previous section title logic, return byteOffsetStart objects
Per patch feedback for T234096 & T236189, this is an alternate approach for receiving
section information for diffs. Instead of wikidiff2 being responsible for determining
sections, this returns byteOffsetStart values for both the from text and to text.
Externally users can determine their own section ranges, compare them to the
byteOffsetStart outputs of each diff item and determine which section a diff item belongs to.

Bug: T237445
Change-Id: Id404f966cd0de92f4a947bf609249ae746ef3326