Use StringStream for Wikidiff2::result
Instead of appending to a string, occasionally with help from a
temporary StringStream, just use a StringStream throughout. This makes
it easy to add integers to the buffer, avoiding scary comments like
"should be plenty" (when calling snprintf). And it should be faster
since it avoids string binary plus, which allocates a temporary string.

Also,

* Replace needsJSONFormat() with new virtual methods printFileHeader()
  and printFileFooter(), avoiding awareness of JSON in Wikidiff2.cpp.
* Rearrange string construction in InlineDiffJSON to avoid temporary
  strings.
* Fix indenting issue.

Change-Id: Ia03641ad2cc4d4f2dab1be14182d0e7f3feaef2c