Split out LineDiffProcessor and add WordDiffCache
Specialization of DiffEngine<Word>/DiffEngine<String> based on whatever
we need out of word and line diffs seemed inelegant. I needed somewhere
else to put post-processing of the line diff. So I made
LineDiffProcessor.

Word diffs were being constructed in a lot of different places, leading
to duplicate work. So I added a simple cache. There's no limit to the
cache size, but memory usage seems acceptable for real test cases.

printMovedLineDiff() was doing diffs of one line against various other
lines. The same word vector is needed for multiple diffs. So we need a
cache of the explodeWords() output. This is especially impactful for
Thai.

Previously, the WordDiffStats diff in looksLikeChange() was done with no
complexity limit, because the line diff was done with no complexity
limit and the limit was passed through. Now that we are using the
correct complexity limit, charSimilarity cannot be correctly populated.
Treating a bailout as having high similarity preserves the existing test
output, formatting bailouts as a changed line instead of delete+add.

Benchmark before:

[en] best:  27.71ms   worst:  28.62ms   avg:  28.19ms
[hi] best:  31.95ms   worst:  33.51ms   avg:  32.30ms
[ru] best:  52.03ms   worst:  53.19ms   avg:  52.60ms
[th] best:  59.86ms   worst:  69.06ms   avg:  60.83ms
[zh] best:  19.78ms   worst:  20.31ms   avg:  20.00ms

After:

[en] best:  19.55ms   worst:  20.66ms   avg:  20.01ms
[hi] best:  21.17ms   worst:  24.68ms   avg:  21.63ms
[ru] best:  37.73ms   worst:  38.83ms   avg:  38.19ms
[th] best:  25.31ms   worst:  31.15ms   avg:  26.15ms
[zh] best:  14.86ms   worst:  15.64ms   avg:  15.17ms

Bug: T324803
Change-Id: Ifd1908f85c75cc39938d99848567c3cd9606b0af